+++ CHANGELOG.md
@@ -13,8 +13,12 @@
     Click to see more.
   </summary>
 
+  ### :bug: Bug Fix
+  - `editor-common`
+    - [#560](https://github.com/wix-incubator/rich-content/pull/560) getBlockInfo returns entity type rather block type
   - `link`
     - [#546](https://github.com/wix-incubator/rich-content/pull/546) saves the last data and the initial state of the checkboxes("Open in a new tab", "Add a nofollow tag") is according to the defaults (anchorTarget, relValue)
+
 </details>
 <hr/>